Code-refactoring

+++ services/members.service.js
@@ -5,25 +5,27 @@ const dayjs = require('dayjs');
 const sendMail = require('../mail');
 const MemberRepository = require('../repositories/members.repository');
 const AWS = require('aws-sdk');
-const { CustomError } = require('../customClass');
+const { CustomError, ServiceReturn } = require('../customClass');
 
 class MemberService {
     memberRepository = new MemberRepository();
+
+    //** 회원가입 */
     signUp = async ({ email, nickname, password, name, phone, address, authCode, url }) => {
         const overlapEmail = await this.memberRepository.findOne({ email });
-        if (overlapEmail) throw { code: 405, result: '이미 사용중인 이메일 입니다.' };
+        if (overlapEmail) throw new CustomError('이미 사용중인 이메일 입니다.', 403);
 
         const overlapNickname = overlapEmail?.nickname == nickname;
-        if (overlapNickname) throw { code: 405, result: '이미 사용중인 닉네임 입니다.' };
+        if (overlapNickname) throw new CustomError('이미 사용중인 닉네임 입니다.', 403);
 
         const isEmailValid = await this.memberRepository.findOneIsEmailValid({ email });
-        if (!isEmailValid) throw { code: 405, result: '이메일을 인증해 주세요.' };
+        if (!isEmailValid) throw new CustomError('이메일을 인증해 주세요.', 402);
 
         const isEmailValidauthCode = isEmailValid?.auth_code == authCode;
-        if (!isEmailValidauthCode) throw { code: 405, result: '인증번호가 일치하지 않습니다.' };
+        if (!isEmailValidauthCode) throw new CustomError('인증번호가 일치하지 않습니다.', 401);
 
         const isEmailValidExpiryTime = dayjs().diff(new Date(isEmailValid.created_at), 'm') >= 30;
-        if (isEmailValidExpiryTime) throw { code: 405, result: '이메일 인증 시간이 초과되었습니다.\n이메일 인증을 재시도 해주세요.' };
+        if (isEmailValidExpiryTime) throw new CustomError('이메일 인증 시간이 초과되었습니다.\n이메일 인증을 재시도 해주세요.', 408);
 
         const passwordToCrypto = crypto.pbkdf2Sync(password, SECRET_KEY.toString('hex'), 11524, 64, 'sha512').toString('hex');
         const expiryDate = dayjs().add(1, 'year').endOf('day').$d;
@@ -31,12 +33,13 @@ class MemberService {
 
         await this.memberRepository.createMember({ email, nickname, passwordToCrypto, group, name, phone, address, expiryDate });
 
-        return { code: 201, result: '회원가입이 완료되었습니다.' };
+        return new ServiceReturn('회원가입이 완료되었습니다.', 201, true);
     };
 
+    //** 회원가입 이메일 인증 */
     isEmailValid = async ({ email }) => {
         const overlapEmail = await this.memberRepository.findOne({ email });
-        if (overlapEmail) throw { code: 405, result: '이미 사용중인 이메일 입니다.' };
+        if (overlapEmail) throw new CustomError('이미 사용중인 이메일 입니다.', 403);
 
         const authCode = String(Math.floor(Math.random() * 1000000)).padStart(6, '0');
         await this.memberRepository.createIsEmailValid({ email, authCode });
@@ -51,51 +54,52 @@ class MemberService {
             console.error('[이메일 발송 실패]', err);
         }
 
-        return { code: 200, result: '인증번호가 발송되었습니다.\n이메일을 확인해 주세요.' };
+        return new ServiceReturn('인증번호가 발송되었습니다.\n이메일을 확인해 주세요.', 200, true);
     };
 
+    //** 로그인 */
     login = async ({ email, password, url }) => {
         const passwordToCrypto = crypto.pbkdf2Sync(password, SECRET_KEY.toString('hex'), 11524, 64, 'sha512').toString('hex');
         const group = url == '/user/login' ? 0 : 1;
         const findUser = await this.memberRepository.findOne({ email });
-        if (!findUser) throw { code: 401, result: '가입되지 않은 이메일 입니다.' };
+        if (!findUser) throw new CustomError('가입되지 않은 이메일 입니다.', 401);
 
         const groupValid = findUser?.group == group;
-        if (!groupValid) throw { code: 401, result: '가입되지 않은 이메일 입니다.' };
+        if (!groupValid) throw new CustomError('가입되지 않은 이메일 입니다.', 401);
 
         const passwordValid = findUser?.password == passwordToCrypto;
-        if (!passwordValid) throw { code: 401, result: '패스워드가 일치하지 않습니다.' };
+        if (!passwordValid) throw new CustomError('패스워드가 일치하지 않습니다.', 401);
 
-        const payload = {
-            member_id: findUser.member_id,
-        };
+        const payload = { member_id: findUser.member_id };
 
-        return { code: 200, result: '로그인 성공', payload };
+        return new ServiceReturn('로그인에 성공하였습니다', 200, payload);
     };
 
+    //** 로그아웃 */
     logout = async () => {
-        return { code: 200, result: '로그아웃 성공' };
+        return new ServiceReturn('정상 로그아웃 되었습니다.', 200, true);
     };
 
+    //** 회원 정보 수정 */
     updateMember = async (member_id, name, nickname, address, phone, image) => {
         await this.memberRepository.updateMember(member_id, nickname, name, phone, address, image);
-        return { code: 200, result: '회원 정보를 수정하였습니다.' };
+        return new ServiceReturn('회원 정보를 수정하였습니다.', 200, true);
     };
 
-    updatePassword = async (member_id, password, changePwd, confirmPwd) => {
-        let passwordToCrypto = crypto.pbkdf2Sync(password, SECRET_KEY.toString('hex'), 11524, 64, 'sha512').toString('hex');
+    //** 회원 비밀번호 변경 */
+    updatePassword = async (member_id, password, changePwd) => {
+        const passwordToCrypto = crypto.pbkdf2Sync(password, SECRET_KEY.toString('hex'), 11524, 64, 'sha512').toString('hex');
 
-        if (changePwd || confirmPwd) {
-            if (changePwd == confirmPwd) {
-                const changePwdToCrypto = crypto.pbkdf2Sync(changePwd, SECRET_KEY.toString('hex'), 11524, 64, 'sha512').toString('hex');
-                passwordToCrypto = changePwdToCrypto;
-            }
-        }
-        await this.memberRepository.updatePassword(member_id, passwordToCrypto);
+        const currentPasswordIsValid = await this.memberRepository.findOne({ password: passwordToCrypto });
+        if (!currentPasswordIsValid) throw new CustomError('현재 암호가 일치하지 않습니다.', 401);
 
-        return { code: 200, result: '비밀번호를 수정하였습니다.' };
+        const changePwdToCrypto = crypto.pbkdf2Sync(changePwd, SECRET_KEY.toString('hex'), 11524, 64, 'sha512').toString('hex');
+        await this.memberRepository.updatePassword(member_id, changePwdToCrypto);
+
+        return new ServiceReturn('패스워드 변경이 완료되었습니다.', 200, true);
     };
 
+    //** 회원 정보 불러오기 */
     findMember = async (member_id) => {
         const findMember = await this.memberRepository.findOne({ member_id });
 
@@ -111,34 +115,27 @@ class MemberService {
             created_at: findMember.created_at,
             updated_at: findMember.updated_at,
         };
-        return { code: 200, member };
+        return new ServiceReturn('회원 정보를 정상적으로 전달하였습니다', 200, member);
     };
 
+    //** 회원 탈퇴 */
     deleteMember = async (member_id_session, member_id, password) => {
-        const findMember = await this.memberRepository.findOne({ member_id });
         const passwordToCrypto = crypto.pbkdf2Sync(password, SECRET_KEY.toString('hex'), 11524, 64, 'sha512').toString('hex');
+        const findMember = await this.memberRepository.findOne([{ member_id, password: passwordToCrypto }]);
 
-        const error = new Error();
-        if (member_id_session != findMember.member_id) {
-            error.message = '회원 탈퇴 권한이 없습니다.';
-            error.status = 403;
-            throw error;
-        } else if (passwordToCrypto != findMember.password) {
-            error.message = '비밀번호가 일치하지 않습니다.';
-            error.status = 412;
-            throw error;
-        }
-
+        if (!findMember) throw new CustomError('패스워드가 일치하지 않습니다.', 401);
         await this.memberRepository.deleteMember(member_id);
 
-        return { code: 200, result: '회원 탈퇴에 성공하였습니다.' };
+        return ServiceReturn('회원 탈퇴에 성공하였습니다.', 200, true);
     };
 
+    //** 회원 프로필 사진 추가 */
     updateProfileImage = async ({ member_id, image }) => {
         await this.memberRepository.updateProfileImage({ member_id, image });
-        return { code: 200, result: '프로필 사진이 정상 저장되었습니다.' };
+        return ServiceReturn('프로필 사진이 정상 저장되었습니다.', 200, true);
     };
 
+    //** 회원 프로필 사진 삭제 */
     deleteProfileImage = async ({ member_id }) => {
         const findUser = await this.memberRepository.findOne({ member_id: member_id });
         const imageKey = findUser.image.replace('https://toydeliverycloud.s3.ap-northeast-2.amazonaws.com/', '');
@@ -151,12 +148,12 @@ class MemberService {
                 Key: imageKey,
             },
             async (err) => {
-                if (err) throw new CustomError('삭제에 실패하였습니다.', 406);
+                if (err) throw new CustomError('삭제에 실패하였습니다.', 500);
                 await this.memberRepository.updateProfileImage({ member_id, image: null });
             }
         );
 
-        return { code: 200, result: '정상적으로 삭제되었습니다.' };
+        return ServiceReturn('정상적으로 삭제되었습니다.', 200, true);
     };
 }
